// packages/luna-core/src/type.d.ts
export type Sections = Array<Section>;
export type Value = string | number | readonly string[];
export type Option = {
    description?: string;
    label: string;
    value: string;
};
export type List = {
    advanced?: {
        action?: string;
        length?: Length<number>;
    };
    description?: string;
    fields: Array<Field | Column>;
    label?: string;
    name: string;
    type: 'list' | (string & {});
} & Base;
export type Fields = Array<Field | Column | List>;
export type Base = Orderable & Hideable;
export type CommonProps = {
    disabled?: boolean;
    id?: string;
    name?: string;
    placeholder?: string;
    required?: boolean;
};
export type Section = {
    advanced?: boolean;
    compact?: boolean;
    description?: string;
    fields?: Fields;
    id?: string | number;
    separator?: boolean;
    title?: string;
} & Base;